Package updates.
A very strange behavior happened when the fhirserver project validated a bearer token.  It would call to .well-known/openid-configuration and would always serialize the jwks_uri as a null entry in the discoveryDocument.  When I removed the secrets that were connected to an entry like this "Kestrel:Certificates:Development:Password": "****"
I believe I need this when including my own ssl cert, not sure.  But removing it solved my issue.